Adds department to Roles Array for clarity
